% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.r
\name{app_sys}
\alias{app_sys}
\title{Access files in the current app}
\usage{
app_sys(...)
}
\arguments{
\item{...}{Path to the file, not starting with /}
}
\value{
The full path to the file
}
\description{
Access files in the current app
}
