% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_upset.r
\name{create_zip_file}
\alias{create_zip_file}
\title{Create ZIP file of all intersection files}
\usage{
create_zip_file(export_dir)
}
\arguments{
\item{export_dir}{Directory containing files to zip}
}
\value{
Path to the created zip file
}
\description{
Create ZIP file of all intersection files
}
