% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.r
\name{get_golem_config}
\alias{get_golem_config}
\title{The application configuration}
\usage{
get_golem_config(
  value = NULL,
  config = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  use_parent = TRUE
)

get_golem_config(
  value = NULL,
  config = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  use_parent = TRUE
)
}
\arguments{
\item{value}{The value to retrieve from the config}

\item{config}{The name of the configuration to use}

\item{use_parent}{If TRUE, the parent config will be used if the value is not found}

\item{...}{Additional parameters passed to `golem::get_golem_config()`}
}
\value{
The configuration of the application

The value from the config
}
\description{
The application configuration

Read App Config
}
