% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_upset.r
\name{make_upset_plot_from_data}
\alias{make_upset_plot_from_data}
\title{Generate UpSet Plot from data}
\usage{
make_upset_plot_from_data(
  data_df,
  id_col,
  group_col,
  plot_title = "UpSet Plot",
  color = "darkblue",
  remove_duplicates = TRUE,
  export_dir = "intersection_sets"
)
}
\arguments{
\item{data_df}{A data frame containing the data}

\item{id_col}{The name of the identifier column}

\item{group_col}{The name of the grouping column}

\item{plot_title}{The title for the plot}

\item{color}{The color for the plot}

\item{remove_duplicates}{Whether to remove duplicate pairs}

\item{export_dir}{Directory to export intersection sets}
}
\value{
A list containing the plot and export information
}
\description{
Generate UpSet Plot from data
}
