% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.r
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(onStart = NULL, options = list(), enableBookmarking = NULL, ...)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.
This is typically used for shinyApp().}

\item{options}{Named options that should be passed to the `shinyApp()` call.}

\item{enableBookmarking}{Can be "url" or "server", the same as in shinyApp()}

\item{...}{arguments to pass to golem_opts.
See `?golem::get_golem_options` for more details.}
}
\description{
Run the Shiny Application
}
